+++ CHANGELOG.md
@@ -7,8 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org).
 
 
-[unreleased]
-------------
+[v0.3.1]
+--------
 
 ### Added
 - Ability to specify a RefreshJob class name